Update return codes in all progs so that failure returns code >= 10
Also made return codes > 0 even if already != 0.
This is because bash return codes are % 255, so -1 will be
returned as 255, making it impossible for the user to then
search for the relevant bit of the code.

I have started return codes in each program at 10, and
each different error has the next integer.
This is because error code 1 is so general that it is not
useful, and 2 is reserved already.